url_checker/requirements.txt
Django >= 5.0
djangorestframework
requests